Adding a new helper funtion to check and output clustering and move other helper fuunctions to the utils file
