+++ vpr/src/pack/cluster_util.h
@@ -55,20 +55,27 @@ struct t_molecule_stats {
     int num_used_ext_outputs = 0; //Number of *used external* output pins across all primitives in molecule
 };
 
+void check_clustering();
 
 //calculate the initial timing at the start of packing stage
-void calc_init_packing_timing (const t_packer_opts& packer_opts, 
+void calc_init_packing_timing (const t_packer_opts& packer_opts,
 							   const t_analysis_opts& analysis_opts,
-							   const std::unordered_map<AtomBlockId, t_pb_graph_node*>& expected_lowest_cost_pb_gnode, 
-							   std::shared_ptr<PreClusterDelayCalculator>& clustering_delay_calc, 
-							   std::shared_ptr<SetupTimingInfo>& timing_info, 
+							   const std::unordered_map<AtomBlockId, t_pb_graph_node*>& expected_lowest_cost_pb_gnode,
+							   std::shared_ptr<PreClusterDelayCalculator>& clustering_delay_calc,
+							   std::shared_ptr<SetupTimingInfo>& timing_info,
 							   vtr::vector<AtomBlockId, float>& atom_criticality);
 
 //Free the clustering data structures
-void free_clustering_data(const t_packer_opts& packer_opts, 
-						  vtr::vector<ClusterBlockId, std::vector<t_intra_lb_net>*>& intra_lb_routing, 
-						  int* hill_climbing_inputs_avail, 
-						  t_cluster_placement_stats* cluster_placement_stats, 
-						  t_molecule_link* unclustered_list_head, 
-						  t_molecule_link* memory_pool, 
+void free_clustering_data(const t_packer_opts& packer_opts,
+						  vtr::vector<ClusterBlockId, std::vector<t_intra_lb_net>*>& intra_lb_routing,
+						  int* hill_climbing_inputs_avail,
+						  t_cluster_placement_stats* cluster_placement_stats,
+						  t_molecule_link* unclustered_list_head,
+						  t_molecule_link* memory_pool,
 						  t_pb_graph_node** primitives_list);
+
+void check_and_output_clustering(const t_packer_opts& packer_opts,
+	  							 const std::unordered_set<AtomNetId>& is_clock,
+	  							 const t_arch* arch,
+	  							 const int& num_clb,
+	  							 const vtr::vector<ClusterBlockId, std::vector<t_intra_lb_net>*>& intra_lb_routing);
